Add Repository, Change some logic, Add Services & DI

+++ Game/BuisnessLogic/SomeLogic.cs
@@ -7,13 +7,13 @@ namespace Game
     public class SomeLogic
     {
         Random random = new Random();
-        public float MakeDamage(float basicDamage, float critChance /*0-100*/, float critPower/*/100+*/)
+        public double MakeDamage(double basicDamage, double critChance /*0-100*/, double critPower/*/100+*/)
         {
             if (random.Next(0, 101) <= critChance)
                 return basicDamage / 100 * critPower; //crit damage
             return basicDamage; //basic non crit damage
         }
-        public float TakeDamage(float damage, float vortex /*0-100*/, float protection/*0-100*/)
+        public double TakeDamage(double damage, double vortex /*0-100*/, double protection/*0-100*/)
         {
             if (random.Next(0, 101) <= vortex)
                 return 0; //miss
